migrationファイル修正、各自でmigrate実行してください

+++ mamaiko/database/migrations/2021_05_21_005447_create_hotels_table.php
@@ -16,7 +16,7 @@ class CreateHotelsTable extends Migration
         Schema::create('hotels', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name'); //ホテル名
-            $table->integer('category_id')->unsigned()->index(); //分類コード、エラーが起きたら真っ先に見直す
+            $table->bigInteger('category_id')->unsigned()->index(); //分類コード、エラーが起きたら真っ先に見直す
             $table->string('address'); //住所
             $table->string('access'); //アクセス
             $table->string('description'); //紹介文
